+++ docs/zh-cn/changelog.md
@@ -19,7 +19,7 @@
 - **Bug Fix**
 
   - 🐞 修复了在iframe沙箱下循环嵌套的问题。
-  - 🐞 修复了在iframe沙箱下`iconfont.js`报错导致图标渲染失败的问题。
+  - 🐞 修复了在iframe沙箱下开启`inline`模式导致通过`getElementsByTagName`获取script元素失败的问题。
 
 - **Update**